refactor(mcp-server): SMI-5949 Wave 2 Step 1 -- extract registry-tools.schemas.ts
Extracts privateRegistryPublishToolSchema/privateRegistryManageToolSchema
(492 -> 434 lines) into a new companion file, following the established
.types.ts/.action.ts convention. Frees line budget for the credential-swap
and review-action steps that follow. index.ts needs no change -- both
schemas are re-exported from registry-tools.ts unchanged.

Refs SMI-5949
